KeySound/requirements.txt
pynput>=1.7.6
customtkinter>=5.2.2
pystray>=0.19.5
Pillow>=10.3.0
yt-dlp>=2024.1.0
python-vlc>=3.0.20123
pytest>=8.2.0
